+++ src/grscheller/fp/nada.py
@@ -12,7 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""### An attempt to give Python a "bottom" type
+"""### An attempt to give Python a "bottom" type.
+
+##### Missing value types.
+
+* **nada:** An attempt to give Python a "bottom" type
 
 While a true bottom type has no instances, `nada` is a singleton. Python's
 evolving typing system seems to reject the concept of a true bottom type.